refactor(notification): webhook handler
