+++ app/common/notification.go
@@ -12,43 +12,33 @@ import (
 	notificationadapter "github.com/openmeterio/openmeter/openmeter/notification/adapter"
 	notificationservice "github.com/openmeterio/openmeter/openmeter/notification/service"
 	notificationwebhook "github.com/openmeterio/openmeter/openmeter/notification/webhook"
+	webhooknoop "github.com/openmeterio/openmeter/openmeter/notification/webhook/noop"
+	webhooksvix "github.com/openmeterio/openmeter/openmeter/notification/webhook/svix"
 	"github.com/openmeterio/openmeter/openmeter/productcatalog/feature"
 )
 
 var Notification = wire.NewSet(
 	NewNotificationService,
+	NewNotificationWebhookHandler,
 )
 
 func NewNotificationService(
 	logger *slog.Logger,
 	db *entdb.Client,
-	notificationConfig config.NotificationConfiguration,
-	svixConfig config.SvixConfig,
+	webhook notificationwebhook.Handler,
 	featureConnector feature.FeatureConnector,
 ) (notification.Service, error) {
-	var notificationRepo notification.Repository
-	notificationRepo, err := notificationadapter.New(notificationadapter.Config{
+	adapter, err := notificationadapter.New(notificationadapter.Config{
 		Client: db,
-		Logger: logger.WithGroup("notification.postgres"),
+		Logger: logger,
 	})
 	if err != nil {
-		return nil, fmt.Errorf("failed to initialize notification repository: %w", err)
-	}
-
-	var notificationWebhook notificationwebhook.Handler
-	notificationWebhook, err = notificationwebhook.New(notificationwebhook.Config{
-		SvixConfig:              svixConfig,
-		RegistrationTimeout:     notificationConfig.Webhook.EventTypeRegistrationTimeout,
-		SkipRegistrationOnError: notificationConfig.Webhook.SkipEventTypeRegistrationOnError,
-		Logger:                  logger.WithGroup("notification.webhook"),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to initialize notification webhook handler: %w", err)
+		return nil, fmt.Errorf("failed to initialize notification adapter: %w", err)
 	}
 
 	notificationService, err := notificationservice.New(notificationservice.Config{
-		Repository:       notificationRepo,
-		Webhook:          notificationWebhook,
+		Adapter:          adapter,
+		Webhook:          webhook,
 		FeatureConnector: featureConnector,
 		Logger:           logger.With(slog.String("subsystem", "notification")),
 	})
@@ -57,4 +47,27 @@ func NewNotificationService(
 	}
 
 	return notificationService, nil
+}
+
+func NewNotificationWebhookHandler(
+	logger *slog.Logger,
+	webhookConfig config.WebhookConfiguration,
+	svixConfig config.SvixConfig,
+) (notificationwebhook.Handler, error) {
+	if !svixConfig.IsEnabled() {
+		return webhooknoop.New(logger), nil
+	}
+
+	handler, err := webhooksvix.New(webhooksvix.Config{
+		SvixConfig:              svixConfig,
+		RegisterEventTypes:      notificationwebhook.NotificationEventTypes,
+		RegistrationTimeout:     webhookConfig.EventTypeRegistrationTimeout,
+		SkipRegistrationOnError: webhookConfig.SkipEventTypeRegistrationOnError,
+		Logger:                  logger.WithGroup("notification.webhook"),
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize notification webhook handler: %w", err)
+	}
+
+	return handler, nil
 }
